+++ 031_image_text/main.c
@@ -27,7 +27,7 @@ int main(void) {
     Texture2D texture = LoadTextureFromImage(parrots); // Image converted to texture, uploaded to VRAM
     UnloadImage(parrots); // Once image has been converted to texture and uploaded to VRAM, it can be unloaded from RAM
 
-    Vector2 position = { (float)(screenWidth / 2 - texture.width / 2), (float)(screenHeight / 2 - texture.height / 2 - 20) };
+    Vector2 position = { ((float)screenWidth / 2 - (float)texture.width / 2), ((float)screenHeight / 2 - (float)texture.height / 2 - 20) };
 
     bool showFont = false;
